requirements.txt
python-dotenv
pydantic
fastapi
uvicorn[standard]

# LLM Providers
openai
anthropic
huggingface_hub

# Pipeline
langgraph
langsmith

# Services
firecrawl-py
supabase
